requirements.txt
pytest~=8.3.4
allure-python-commons~=2.13.5
requests~=2.32.3
websockets~=14.2
aiohttp~=3.11.12